+++ packages/specflow/src/lib/init-cancelled.ts
@@ -1,0 +1,6 @@
+export class InitCancelledError extends Error {
+  constructor() {
+    super("INIT_CANCELLED");
+    this.name = "InitCancelledError";
+  }
+}